+++ web-dashboard/src/scss/SiteAnalytics.scss
@@ -1,10 +1,10 @@
-.station-list {
+.site-list {
     list-style: none;
     padding: 0;
     margin: 0;
 }
 
-.station-list-item {
+.site-list-item {
     display: flex;
     justify-content: space-between;
     align-items: center;
@@ -29,13 +29,13 @@
         margin-right: 10px;
     }
 
-    // Styling for the station information
-    .station-info {
+    // Styling for the site information
+    .site-info {
         // Add any additional styling you want for the text
     }
 
     // Styling for the buttons
-    .station-actions {
+    .site-actions {
         button {
             margin-left: 10px; // Space between buttons
         }
@@ -59,8 +59,8 @@
     }
 }
 
-// StationAnalytics.scss
-.station-analytics-container {
+// siteAnalytics.scss
+.site-analytics-container {
     display: flex;
     flex-direction: column;
     align-items: center;
